# account-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  account-service_internal_model.AccountResponse:
    properties:
      account_number:
        example: "1234567890"
        type: string
      balance:
        example: 1000.5
        type: number
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  account-service_internal_model.BalanceResponse:
    properties:
      account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      account_number:
        example: "1234567890"
        type: string
      balance:
        example: 1000.5
        type: number
    type: object
  account-service_internal_model.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
  internal_handler.HealthResponse:
    properties:
      service:
        example: account-service
        type: string
      status:
        example: ok
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: Bank account management service for Minibank
  title: Minibank Account Service API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Retrieve all bank accounts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/account-service_internal_model.AccountResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my accounts
      tags:
      - accounts
    post:
      description: Creates a new bank account for the authenticated user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account-service_internal_model.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new bank account
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: Retrieve a bank account by its UUID (must be owned by authenticated
        user)
      parameters:
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account-service_internal_model.AccountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
  /accounts/{id}/balance:
    get:
      description: Retrieve the balance of a specific bank account owned by the authenticated
        user
      parameters:
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account-service_internal_model.BalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/account-service_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account balance
      tags:
      - accounts
  /health:
    get:
      description: Returns service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.HealthResponse'
      summary: Health check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Enter your token with the Bearer prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiIs..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"